joinform coming along nicely - checkpoint
